+++ components/EpisodeItem.vue
@@ -1,5 +1,4 @@
 <script setup>
-import { useToast } from "primevue/usetoast"
 import VImage from "@nypublicradio/nypr-design-system-vue3/v2/src/components/VImage.vue"
 // TEMP fix to make ripple work
 import { usePrimeVue } from "primevue/config"
@@ -8,10 +7,7 @@ import DownloadIcon from "~/components/icons/DownloadIcon.vue"
 import ShareIcon from "~/components/icons/ShareIcon.vue"
 //import QueueIcon from "~/components/icons/QueueIcon.vue"
 import {
-  deleteFavorite,
-  saveFavorite,
   checkIsFavorited,
-  getFavoritedItems,
   trackClickEvent,
   shareAPI,
   getMinutes,
@@ -19,8 +15,9 @@ import {
   fetchDuration,
   hasAudio,
   togglePlayEpisode,
+  addToFavorites,
 } from "~/utilities/helpers"
-import { useCurrentUser, useAccountPromptSideBar } from "~/composables/states"
+import { useCurrentUser } from "~/composables/states"
 import {
   fetchAndStoreMp3,
   getDownloadedImageUri,
@@ -29,7 +26,6 @@ import {
   /*   formatFileSize, */
 } from "~/utilities/file-system"
 import { FALLBACKIMAGEEP } from "~/composables/globals"
-const toast = useToast()
 
 const $primevue = usePrimeVue()
 defineExpose({
@@ -68,7 +64,7 @@ const props = defineProps({
     default: false,
   },
 })
-const accountPromptSideBar = useAccountPromptSideBar()
+//const accountPromptSideBar = useAccountPromptSideBar()
 const user = useCurrentUser()
 
 // check if item is already favorited
@@ -96,37 +92,16 @@ watch(
 //   trackClickEvent("Click Tracking - Add to Queue", "Episode Item", bucketItem.title)
 // }
 
-const handleAddToFavorites = async (bucketItem) => {
+// add item to favorites
+const handleAddToFavorites = (bucketItem) => {
+  // helper func for adding to favorites, also handles account prompt if not logged in
+  addToFavorites(bucketItem, isFavorited.value)
   if (user.value) {
-    const episode = {
-      ...bucketItem,
-      slug: bucketItem.meta?.slug ?? bucketItem.slug,
-    }
-    if (isFavorited.value) {
-      await deleteFavorite(episode)
-      getFavoritedItems()
-      isFavorited.value = false
-      emit("onDeleteFavorite")
-    } else {
-      await saveFavorite(episode, episode.type)
-      getFavoritedItems()
-      isFavorited.value = true
-      emit("onSaveFavorite")
-    }
-    toast.add({
-      severity: "info",
-      summary: "Updated your favorites.",
-      life: 3000,
-    })
-    trackClickEvent(
-      "Click Tracking - Add/remove from favorites",
-      "Episode Item",
-      bucketItem.title
-    )
-  } else {
-    accountPromptSideBar.value = true
+    isFavorited.value = !isFavorited.value
+    isFavorited.value ? emit("onDeleteFavorite") : emit("onSaveFavorite")
   }
 }
+
 const progress = ref(null)
 // handle the download of the audio file request and feed the progress
 const handleDownload = async (bucketItem) => {
